pyproject.toml
[project]
name = "examen-bentoml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bentoml>=1.4.35",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "pydantic==2.11.7",
    "pyjwt>=2.11.0",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
